Rename file and folder names.
