+++ src/entity/pokerboardInvite/model.ts
