+++ src/entity/userTicket/model.ts
