+++ src/helpers/user.helper.ts
@@ -1,5 +1,5 @@
 import { ErrorMessages } from '../constants/message';
-import { registrationSchema } from '../schemas/user_schema';
+import { registrationSchema } from '../schemas/user.schema';
 import { validateData } from '../utils/common';
 
 export const validateUserRegistrationData = (data: object) => {
