+++ src/index.ts
@@ -4,7 +4,7 @@ import express from 'express';
 
 import { LogMessages } from './constants/message';
 import { connectToMySQLDB } from './database/mysql';
-import userRouter from './routes/user_routes';
+import userRouter from './routes/user.route';
 import config from './settings/config';
 import { Routes } from './constants/enums';
