+++ src/routes/user.route.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 
-import { registerUser } from '../controllers/user_registration';
-import { registerUserValidation } from '../middlewares/user_validation';
+import { registerUser } from '../controllers/user.controller.registration';
+import { registerUserValidation } from '../middlewares/user.middleware';
 
 const router: express.Router = express.Router();
